requirements.txt
pandas
openai
python-dotenv
tqdm
